app/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
greenlet>=3.1.1
python-dotenv==1.0.1
pydantic-settings==2.7.0
alembic==1.14.0
httpx==0.28.1
gigachat==0.1.43
python-multipart==0.0.21
livekit-api>=1.0.0
redis>=5.0.0
grpcio>=1.60.0
PyJWT>=2.8.0
certifi>=2024.0.0